Added GA_ArcadeItem_HealthPotion
This potion regains health based on a percentage of total health.
It is not working as my vision currently. Just made it for testing.
Will eventually need to not regain charges automatically, so will have to create its own base WithCharges class that instead of checking for cooldown tags it --gameplaytagstacks and ++Gameplaytagstacks when either picked up/granted another way.
Currently has a secondary cooldown for this exact reason, it will not be needed after changes are made.