Refactored PowerUpComponent for ease of use.
It was too much work adding the powerups to a list in a adata asset, now we create a data asset for each power up ( which does not need to have a different one for each quality).
Each quality now supports different abilities/gameplayeffects/tagstacks, this allows for more customization in the powerups.
For example: We can have a Normal quality power up that just gives the ability, but as we progress in qualities, we can start adding effects and tags (more projectiles or an increase it health/damage/mana etc.)